+++ src/index.ts
@@ -1,4 +1,11 @@
 import MongoStore from './lib/MongoStore.js'
+export {
+  createKrupteinAdapter,
+  createWebCryptoAdapter,
+  type CryptoAdapter,
+  type CryptoOptions,
+  type WebCryptoAdapterOptions,
+} from './lib/cryptoAdapters.js'
 
 export default MongoStore
 export { MongoStore }